Wrap all paths in backticks
This seemed to be the standard everywhere else in the document, so do it
for the remaining ones as well.

There were a few links here, such as [./seed/service](./seed/service)
and [CMakeLists.txt](./seed/service/CMakeLists.txt). These don't make
sense, as they don't actually link to anything. I suspect they are
remnants from a time when this document lived in the same repo as
IncludeOS itself?

I also fixed the `CMakelists.txt` path to be to `your_service` instead
of `seed/service`, as that's the one you actually want to update.